Ensure console tests run against postgres
