+++ satellite/console/consoleweb/consoleql/query_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/graphql-go/graphql"
 	"github.com/stretchr/testify/assert"
-	"go.uber.org/zap"
+	"go.uber.org/zap/zaptest"
 
 	"storj.io/storj/internal/testcontext"
 	"storj.io/storj/pkg/auth"
@@ -26,7 +26,7 @@ func TestGraphqlQuery(t *testing.T) {
 		ctx := testcontext.New(t)
 		defer ctx.Cleanup()
 
-		log := zap.NewExample()
+		log := zaptest.NewLogger(t)
 
 		service, err := console.NewService(
 			log,
